minor fixes + move repository commands to commands classes
