--- src/main/java/com/au/mit/vcs/common/commands/CleanCmd.java
@@ -1,26 +1,0 @@
-package com.au.mit.vcs.common.commands;
-
-import com.au.mit.vcs.common.Repository;
-import com.au.mit.vcs.common.exceptions.CommandBuildingException;
-import com.au.mit.vcs.common.exceptions.NotEnoughArgumentsException;
-import org.apache.commons.cli.Options;
-
-import java.util.List;
-import java.util.concurrent.Callable;
-
-/**
- * Created by semionn on 23.09.16.
- */
-public class CleanCmd extends Command {
-    public CleanCmd() {
-        super("clean", new Options());
-    }
-
-    @Override
-    public Callable<Void> createTask(Repository repository, CommandArgs commandArgs) throws CommandBuildingException {
-        return () -> {
-            repository.clean();
-            return null;
-        };
-    }
-}
